refactor: select profile screen & profile list
