+++ src/features/multi-profile/ProfileList.hooks.ts
@@ -1,0 +1,95 @@
+import { useEffect, useState } from 'react';
+import { DrawerNavigationProp } from '@react-navigation/drawer';
+
+import { AppException } from '@covid/core/api/ApiServiceErrors';
+import i18n from '@covid/locale/i18n';
+import { offlineService } from '@covid/Services';
+import { Services } from '@covid/provider/services.types';
+import { ICoreService } from '@covid/core/user/UserService';
+import { useInjection } from '@covid/provider/services.hooks';
+import { Profile } from '@covid/components/Collections/ProfileList';
+
+import { ScreenParamList } from '../ScreenParamList';
+import appCoordinator from '../AppCoordinator';
+
+export const useProfileList = (navigation: DrawerNavigationProp<ScreenParamList, 'SelectProfile'>) => {
+  const userService = useInjection<ICoreService>(Services.User);
+
+  const [shouldRefresh, setShouldRefresh] = useState<boolean>(false);
+  const [status, setStatus] = useState<string>('');
+  const [error, setError] = useState<AppException | null>(null);
+  const [isApiError, setIsApiError] = useState<boolean>(false);
+
+  const [profiles, setProfiles] = useState<Profile[]>([]);
+  const [isLoaded, setIsLoaded] = useState<boolean>(false);
+  const [onRetry, setOnRetry] = useState<any>(() => {});
+
+  useEffect(() => {
+    const unsubscribe = navigation.addListener('focus', async () => {
+      if (shouldRefresh) {
+        await listProfiles();
+      }
+    });
+
+    listProfiles().then(() => {
+      setShouldRefresh(true);
+    });
+
+    return unsubscribe;
+  }, []);
+
+  const retryListProfiles = () => {
+    setStatus(i18n.t('errors.status-retrying'));
+    setError(null);
+    setTimeout(() => listProfiles(), offlineService.getRetryDelay());
+  };
+
+  const listProfiles = async () => {
+    setStatus(i18n.t('errors.status-loading'));
+    setError(null);
+
+    try {
+      const profiles = await userService.listPatients();
+
+      if (profiles) {
+        setProfiles(profiles);
+        setIsLoaded(true);
+      }
+    } catch (error) {
+      setError(error);
+    }
+  };
+
+  const profileSelected = async (profileId: string, index: number) => {
+    try {
+      const currentPatient = await userService.getPatientState(profileId);
+      setIsApiError(false);
+      await appCoordinator.profileSelected(index === 0, currentPatient);
+    } catch (error) {
+      setIsApiError(true);
+      setError(error);
+
+      setOnRetry(() => {
+        setStatus(i18n.t('errors.status-retrying'));
+        setError(null);
+
+        setTimeout(() => {
+          setStatus(i18n.t('errors.status-loading'));
+          profileSelected(profileId, index);
+        }, offlineService.getRetryDelay());
+      });
+    }
+  };
+
+  return {
+    status,
+    error,
+    isApiError,
+    isLoaded,
+    onRetry,
+    profiles,
+    retryListProfiles,
+    profileSelected,
+    setIsApiError,
+  };
+};
